Added initial source files.
The code to create and run a crowd_sim object then display the results graphically will be contained in main.
The code to build and control a simulation will be contained in the simulation::simulator module.
The code to define a pedestrian/walker is contained in the simulation::pedestrian module.