Re-wrote collision avoidance code, fixing most issues
Collisions between pedestrians are now far rarer than they used to be, which makes the overall behaviour look much more realistic.
After tweaking the target speed range slightly, the average travel time is now just above 19s, which is very close to the target of 18.6s. The values are within one standard deviation, which is good enough for me.
The collision avoidance code is a complete mess, and I know how I might improve it - starting with constraining all angles between -π and π instead of 0 and 2π.